Split Renderer into separate class.
